Added nebula models based on existing catalog nebulae
This minimizes the file size of the pak by reusing textures that already exist in the data directory. Changing the colors prevents them from being too recognizable. The Red Nebula uses the same model as the Heart Nebula to ensure that it really is red. Also deleted the Tion system backup file.